@@ -18,26 +18,27 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+2
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  toast: ^0.1.5
-  just_audio: ^0.6.12
-  http: ^0.12.2
-  firebase_core: ^0.4.0+9
-  firebase_auth: ^0.16.0
-  google_sign_in: ^4.4.4
-  # enum_to_string: ^2.0.1
+#  toast: ^0.1.5
+  just_audio: ^0.9.7
+  http: ^0.13.3
+  firebase_core: ^1.6.0
+  firebase_auth: ^3.1.0
+  google_sign_in: ^5.0.7
+  flutter_bloc: ^7.2.0
+  equatable: ^2.0.3
   flutter:
     sdk: flutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
   change_app_package_name: ^1.0.0
-  flutter_launcher_icons: ^0.8.1
+  flutter_launcher_icons: ^0.9.2
   flutter_test:
     sdk: flutter
 
