@@ -1,10 +1,10 @@
-import 'package:bag_of_words/res/AppColor.dart';
+import 'package:bag_of_words/res/app_color.dart';
 import 'package:flutter/material.dart';
 
 class TextEntryWidget extends StatelessWidget {
   final Function onClickedAdd;
 
-  const TextEntryWidget({@required this.onClickedAdd, Key key})
+  const TextEntryWidget({required this.onClickedAdd, Key? key})
       : super(key: key);
 
   @override
