@@ -1,43 +1,13 @@
-import 'package:bag_of_words/controller/Application.dart';
-import 'package:bag_of_words/page/Homepage.dart';
-import 'package:bag_of_words/page/LoginPage.dart';
-import 'package:bag_of_words/res/AppColor.dart';
-import 'package:bag_of_words/res/AppString.dart';
+import 'package:bag_of_words/app.dart';
+import 'package:bag_of_words/bloc/app_bloc_observer.dart';
+import 'package:bag_of_words/data/repos/auth/auth_repo.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 
-void main() => runApp(MyApp());
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    SystemChrome.setSystemUIOverlayStyle(
-      SystemUiOverlayStyle(statusBarColor: Colors.transparent),
-    );
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      title: AppString.appName,
-      theme: ThemeData(
-        primaryColor: AppColor.darkBg,
-        visualDensity: VisualDensity.adaptivePlatformDensity,
-        backgroundColor: AppColor.dark,
-      ),
-      home: _getInitialPage(),
-    );
-  }
-
-  Widget _getInitialPage() {
-    return FutureBuilder(
-      future: Application.getInstance().isAuthenticated(),
-      builder: (context, snapshot) {
-        if (snapshot.connectionState != ConnectionState.done) {
-          return Center(child: CircularProgressIndicator());
-        }
-
-        final isAuthenticated = snapshot.data;
-        if (isAuthenticated != null && isAuthenticated) return HomePage();
-        return LoginPage();
-      },
-    );
-  }
+Future<void> main() async {
+  Bloc.observer = AppBlocObserver();
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
+  runApp(App(authRepo: AuthRepo()));
 }
