@@ -1,5 +1,5 @@
 import 'package:bag_of_words/controller/Application.dart';
-import 'package:bag_of_words/res/AppColor.dart';
+import 'package:bag_of_words/res/app_color.dart';
 import 'package:flutter/material.dart';
 
 class HomePageActions extends StatelessWidget {
